+++ CHANGELOG.md
@@ -2,10 +2,14 @@
 
 # UNRELEASED
 
+* EXOATPPolicyForO365
+  * Deprecated properties AllowClickThrough & EnableSafeLinksForO365Clients.
 * EXOAuthenticationPolicyAssignment
   * Initial release.
 * EXOCASMailboxSettings
   * New resource to configure Exchange Online CAS Mailbox settings.
+* EXOSafeLinksPolicy
+  * Deprecated properties DoNotAllowClickThrough, DoNotTrackUserClicks & IsEnabled.
 * IntuneAppProtectionPolicyiOS
    * Fixed issue with creation a new policies where it was complaining about invalid minimum versions.
    * Fixed issues where creating new policies threw an error complaining about an invalid duration format.